+++ BusinessLayer/LogicFacade.cs
@@ -8,24 +8,7 @@ using System;
 
 namespace BeerProductionSystem.BusinessLayer
 {
-    enum Commands
-    {
-        RESET = 1,
-        START,
-        STOP,
-        ABORT,
-        CLEAR
-    }
 
-    enum ProductMaxSpeed  //TODO move out of presentation layer
-    {
-        Pilsner = 600,
-        Wheat = 300,
-        IPA = 150,
-        Stout = 200,
-        Ale = 100,
-        Alcohol_Free = 125
-    }
 
     class LogicFacade : ILogicFacade
     {
